GroupSetting.vueの修正 ①メンバーリストの表示 ②招待可能ユーザーの表示 ③グループ編集機能の作成と表示
※グループ退会機能未完成

+++ app/Image.php
@@ -7,6 +7,8 @@ use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
 use App\User;
+use App\Group;
+use App\Topic;
 
 class Image extends Model
 {
@@ -18,7 +20,7 @@ class Image extends Model
         $data = str_replace('data:'.$mime_type.';base64,', '', $data);
         $data = str_replace(' ', '+', $data);
         $fileData = base64_decode($data);
-        $file_name = 'images/' . Str::random(16) . '.png';
+        $file_name = 'uploads/' . Str::random(16) . '.png';
         Storage::disk('public')->put($file_name, $fileData);
 
         $image = new Image;
@@ -32,17 +34,21 @@ class Image extends Model
     public static function destroy($bool)
     {
         $image_ids = User::whereNotNUll('image_id')->pluck('image_id');
+        $image_ids = $image_ids->merge(Group::whereNotNUll('image_id')->pluck('image_id'));
+        // $image_ids = $image_ids->merge(Topic::whereNotNUll('image_id')->pluck('image_id'));
         $images = Image::whereNotIn('id', $image_ids)->get();
         
         foreach ($images as $image) {
             $name = $image->name;
-            // 画像を消す処理
-            if(strpos($name,'storage/') !== false){
-                $name = str_replace('storage/', '', $name);
-                Storage::disk('public')->delete($name);
+            if (!strpos($name, 'images/')) {
+                // 画像を消す処理
+                if(strpos($name,'storage/') !== false){
+                    $name = str_replace('storage/', '', $name);
+                    Storage::disk('public')->delete($name);
+                }
+                // レコードを消す処理
+                $image->delete();
             }
-            // レコードを消す処理
-            $image->delete();
         }
     }
